Added dependency for amazon s3 integration. Next steps :
pre-signed urls for scalability for upload:
-- check user's paid plan and if they can upload , secure via JWT, and get user info from JWT.
-- for delete , simply use from JWT and file name gotten to delete image from back end .

Important :  cdn shall be omitted for now for this MVP --also change db , or leave empty for now - make it nullable on both fluent api and migration sql file.

Proceed with rest of business cases : and write unit tests to consider this MVP done.